% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candysurvey.R
\docType{data}
\name{candysurvey}
\alias{candysurvey}
\title{2015 Halloween Candy Survey Data}
\format{A data frame with 5658 rows and 103 variables:
\describe{
  \item{user}{user ID}
  \item{trick_or_treat}{boolean -- will you be going trick or treating?}
  \item{archie_girls}{one of \code{Betty}, \code{Veronica}, or \code{???}}
  \item{the_dress}{regarding the dress, is it \code{White and gold} or \code{Blue and black}?}
  \item{imitation}{imitation is a form of: \code{Laziness}, \code{Flattery}, or \code{Lazily flattering someone}}
  \item{friday_or_saturday}{do you prefer \code{Friday} or \code{Sunday}?}
  \item{n_mints}{integer -- how many mints are in my hand?}
  \item{age}{integer -- how old are you?}
  \item{...}{the remaining variable are the candies to which the surveyee responded \code{JOY} or \code{DESPAIR}}
}}
\source{
\url{https://docs.google.com/spreadsheets/d/1REZvjqv0lj3dEYb0CsGyDXkXrjhJ4izlAEImgaufjCc/pubhtml}
}
\usage{
candysurvey
}
\description{
David Ng and Ben Cohen's 2015 Halloween Candy Hierarchy survey data
}
\keyword{datasets}

